# openapi/discovery-v1.yaml
openapi: 3.0.3
info:
  title: CAFE Discovery API
  version: "1.0.0"
  license:
    name: Proprietary
    url: https://example.com/license
  description: |
    Target HTTP contract for CAFE Discovery (observation, scans, CAFE wallets) per **WORKPLAN_API.md**
    §0.1, §2.1–2.2, §4.1–4.2, §5.4, §8.

    **Public URL prefix:** `/api/discovery/v1` (edge). Upstream Discovery may receive paths under
    `/discovery/v1/...` after ingress strips `/api` — see `cafe-deploy` / **WORKPLAN_API_PR.md** PR9.

    **CAFE wallets:** HTTP surface is only under **`/wallets`** in this spec (i.e. **`/api/discovery/v1/wallets/...`** at the edge). Discovery does **not** serve a legacy root **`/wallets`** path without the **`/discovery/v1`** prefix.

    **Companion spec (policies):** `cafe-crypto-policy-mgt/openapi/cpm-v1.yaml` (PR1b); correlation
    key is **`scan_id`** (UUID) for Discovery ↔ CPM Option A.
servers:
  - url: https://{host}/api/discovery/v1
    description: Public edge base (canonical).
    variables:
      host:
        default: api.example.com
        description: API host.
...
security:
  - bearerAuth: []
paths:
...
  /wallets/scans:
    get:
      tags: [wallet-scans]
      summary: List wallet/EVM scans (synopsis)
      description: |
        Paginated synopsis list. **Default sort:** `created_at` descending, then `scan_id`
        descending (deterministic tie-breaker). **`status`** is **lifecycle only**, not crypto posture.

        **`chain_id` without `address` → 400.** Do not use address in path; use query `address`.
      operationId: listWalletScans
      parameters:
        - $ref: "#/components/parameters/QueryAddress"
        - $ref: "#/components/parameters/QueryChainId"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryOffset"
      responses:
        "200":
          description: Paginated list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedScanList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT; owner-scoped resources (WORKPLAN_API.md §4.x).
  parameters:
...
    QueryAddress:
      name: address
      in: query
      required: false
      schema:
        type: string
        description: Normalized Ethereum address filter (wallet scans only).
    QueryChainId:
      name: chain_id
      in: query
      required: false
      schema:
        type: integer
        format: int64
      description: Chain refinement; **only** with `address`.
    QueryLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
    QueryOffset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Unauthorized:
      description: Missing or invalid authentication.
...
    BadRequest:
      description: Invalid request (e.g. mutually exclusive body fields, `chain_id` without `address`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code (e.g. SCAN_REFERENCED_BY_POLICY).
        message:
          type: string
          description: Human-readable message.
    ScanLifecycleStatus:
      type: string
      enum: [requested, started, completed, failed]
      description: Lifecycle metadata only — not crypto posture (WORKPLAN_API.md §5.4.1).
    ScanListItem:
      type: object
      required: [scan_id, created_at, status]
      description: Wallet scan synopsis (WORKPLAN_API.md §5.4.2).
      properties:
        scan_id:
          type: string
          format: uuid
        target_address:
          type: string
          description: Required for scans triggered by address when known.
        chain_ids:
          type: array
          items:
            type: integer
            format: int64
          description: May be empty until chains are known.
        created_at:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/ScanLifecycleStatus"
...
    PaginationEnvelope:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
    PaginatedScanList:
      allOf:
        - $ref: "#/components/schemas/PaginationEnvelope"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ScanListItem"
...